chore: move comments model to administrative templates plugin
