+++ src/plugins/CMakeLists.txt
@@ -21,7 +21,6 @@ add_subdirectory(admx)
 add_subdirectory(cmtl)
 add_subdirectory(cmtx)
 add_subdirectory(common)
-add_subdirectory(comments)
 add_subdirectory(ini)
 add_subdirectory(pol)
 add_subdirectory(preferences)
