+++ src/plugins/administrative_templates/comments/comment.h
