+++ src/plugins/administrative_templates/comments/commentdefinitionresources.h
