+++ src/plugins/administrative_templates/comments/localizationresourcereference.h
