--- src/plugins/comments/CMakeLists.txt
@@ -1,51 +1,0 @@
-find_package(Qt5 COMPONENTS Widgets LinguistTools REQUIRED)
-
-find_package(GPUI COMPONENTS core io REQUIRED)
-
-set(CMAKE_INCLUDE_CURRENT_DIR ON)
-
-set(CMAKE_AUTOMOC ON)
-set(CMAKE_AUTOUIC ON)
-
-set(PLUGIN_NAME comments-plugin)
-
-include_directories("${CMAKE_SOURCE_DIR}/src/core")
-
-set(HEADERS
-    model/comment.h
-    model/commentdefinitionresources.h
-    model/localizationresourcereference.h
-    model/policycomments.h
-    model/policynamespace.h
-    model/policynamespaceassociation.h
-    commentssnapin.h
-)
-
-set(SOURCES
-    commentsplugin.cpp
-    commentssnapin.cpp
-)
-
-set(UI_FORMS
-)
-
-set(SOURCES ${SOURCES} ${HEADERS} ${UI_FORMS})
-
-add_definitions(
-    -DGPUI_COMMENTS_PLUGIN
-)
-
-set(TS_FILES
-)
-
-set_source_files_properties(${TS_FILES} PROPERTIES OUTPUT_LOCATION ${CMAKE_CURRENT_SOURCE_DIR})
-
-qt5_add_translation(QM_FILES ${TS_FILES})
-
-add_translation_resource(RESOURCES_SRC "comments" ${QM_FILES})
-
-qt5_add_resources(LIB_RESOURCES ${RESOURCES_SRC} ${ICON_RESOURCES})
-
-add_gpui_plugin(${PLUGIN_NAME} ${SOURCES} ${LIB_RESOURCES})
-target_link_libraries(${PLUGIN_NAME} Qt5::Core)
-target_link_libraries(${PLUGIN_NAME} ${GPUI_CORE_LIBRARY} ${GPUI_GUI_LIBRARY} ${GPUI_IO_LIBRARY})
